fix restore
