+++ packages/adapter-components/index.ts
@@ -27,6 +27,7 @@ export * from './src/references/element_fixers'
 export * from './src/references/custom_references'
 export * from './src/references/weak_reference_handler'
 export * from './src/resolve_utils'
+export * from './src/restore_utils'
 export * from './src/references/get_references'
 export * from './src/adapter-wrapper'
 export * as openapi from './src/openapi'
