+++ packages/jira-adapter/src/filters/notification_scheme/notification_scheme_deployment.ts
@@ -24,8 +24,8 @@ import {
   isInstanceChange,
   isModificationChange,
 } from '@salto-io/adapter-api'
-import { applyFunctionToChangeData, createSchemeGuard, restoreChangeElement } from '@salto-io/adapter-utils'
-import { resolveValues } from '@salto-io/adapter-components'
+import { applyFunctionToChangeData, createSchemeGuard } from '@salto-io/adapter-utils'
+import { resolveValues, restoreChangeElement } from '@salto-io/adapter-components'
 import { logger } from '@salto-io/logging'
 import { collections } from '@salto-io/lowerdash'
 import { findObject, setFieldDeploymentAnnotations, setTypeDeploymentAnnotations } from '../../utils'
